tech(api): move error message to its appropriate bounded context
