+++ api/src/prescription/campaign/domain/errors.js
@@ -58,6 +58,19 @@ class CampaignParticipationDoesNotBelongToUser extends DomainError {
   }
 }
 
+class OrganizationNotAuthorizedMultipleSendingAssessmentToCreateCampaignError extends DomainError {
+  constructor(organizationId) {
+    const message = `L'organisation ${organizationId}, n'est pas autorisée à créer une campagne d'évaluation à envoi multiple.`;
+    super(message);
+  }
+}
+
+class UserNotAuthorizedToCreateCampaignError extends DomainError {
+  constructor(message = "Cet utilisateur n'est pas autorisé à créer une campagne.") {
+    super(message);
+  }
+}
+
 export {
   ArchivedCampaignError,
   CampaignCodeFormatError,
@@ -66,6 +79,8 @@ export {
   DeletedCampaignError,
   IsForAbsoluteNoviceUpdateError,
   MultipleSendingsUpdateError,
+  OrganizationNotAuthorizedMultipleSendingAssessmentToCreateCampaignError,
   SwapCampaignMismatchOrganizationError,
   UnknownCampaignId,
+  UserNotAuthorizedToCreateCampaignError,
 };
