+++ api/tests/prescription/campaign/unit/application/api/campaigns-api_test.js
@@ -1,6 +1,6 @@
 import * as campaignApi from '../../../../../../src/prescription/campaign/application/api/campaigns-api.js';
+import { UserNotAuthorizedToCreateCampaignError } from '../../../../../../src/prescription/campaign/domain/errors.js';
 import { usecases } from '../../../../../../src/prescription/campaign/domain/usecases/index.js';
-import { UserNotAuthorizedToCreateCampaignError } from '../../../../../../src/shared/domain/errors.js';
 import { Campaign } from '../../../../../../src/shared/domain/models/Campaign.js';
 import { CampaignReport } from '../../../../../../src/shared/domain/read-models/CampaignReport.js';
 import { catchErr, expect, sinon } from '../../../../../test-helper.js';
