+++ api/tests/prescription/campaign/unit/domain/usecases/create-campaign_test.js
@@ -1,7 +1,7 @@
+import { UserNotAuthorizedToCreateCampaignError } from '../../../../../../src/prescription/campaign/domain/errors.js';
 import { CampaignCreator } from '../../../../../../src/prescription/campaign/domain/models/CampaignCreator.js';
 import { createCampaign } from '../../../../../../src/prescription/campaign/domain/usecases/create-campaign.js';
 import { CampaignTypes } from '../../../../../../src/prescription/shared/domain/constants.js';
-import { UserNotAuthorizedToCreateCampaignError } from '../../../../../../src/shared/domain/errors.js';
 import { catchErr, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | create-campaign', function () {
